continue: address pr feedback, enhance authentication UI and internationalization support
