+++ src/app/(auth)/register/page.tsx
@@ -1,29 +1,27 @@
 "use client";
 
 import { useEffect } from "react";
-import { Mail, MoveLeft, ShieldCheck, User } from "lucide-react";
+import { Mail, MoveLeft, User } from "lucide-react";
 import Link from "next/link";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useTranslations } from "next-intl";
 import { Controller } from "react-hook-form";
 
 import { AuthActionLink } from "@/features/auth/components/AuthActionLink";
+import { AuthOtpStep } from "@/features/auth/components/AuthOtpStepCard";
 import { AuthStepHeaderIcon } from "@/features/auth/components/AuthStepHeaderIcon";
 import { ErrorAlert } from "@/features/auth/components/ErrorAlert";
 import { FieldMessageSlot } from "@/features/auth/components/FieldMessageSlot";
 import { GoogleButton } from "@/features/auth/components/GoogleButton";
-import { OtpDevTotpFill } from "@/features/auth/components/OtpDevTotpFill";
 import { StepGuard } from "@/features/auth/components/StepGuard";
 import { StepIndicator } from "@/features/auth/components/StepIndicator";
 import { useRegisterEmailStep } from "@/features/auth/hooks/useRegisterEmailStep";
 import { useRegisterOtpStep } from "@/features/auth/hooks/useRegisterOtpStep";
 import { useRegisterProfileStep } from "@/features/auth/hooks/useRegisterProfileStep";
-import { formatCountdown } from "@/features/auth/utils";
 import { Button } from "@/shared/components/ui/button";
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/shared/components/ui/card";
 import { Field, FieldError, FieldGroup, FieldLabel } from "@/shared/components/ui/field";
 import { Input } from "@/shared/components/ui/input";
-import { InputOTP, InputOTPGroup, InputOTPSlot } from "@/shared/components/ui/input-otp";
 import { Separator } from "@/shared/components/ui/separator";
 
 type RegisterStep = "email" | "otp" | "profile";
@@ -89,7 +87,7 @@ function EmailStep() {
       </CardHeader>
       <CardContent className="space-y-4">
         <ErrorAlert message={apiError.message} />
-        <form onSubmit={onSubmit} className="space-y-4">
+        <form id="register-email-form" onSubmit={onSubmit} className="space-y-4">
           <FieldGroup>
             <Controller
               name="email"
@@ -112,7 +110,7 @@ function EmailStep() {
               )}
             />
           </FieldGroup>
-          <Button type="submit" className="w-full" disabled={form.formState.isSubmitting}>
+          <Button type="submit" form="register-email-form" className="w-full" disabled={form.formState.isSubmitting}>
             {t("register.continue")}
           </Button>
         </form>
@@ -130,85 +128,21 @@ function EmailStep() {
 }
 
 function OtpStep() {
-  const t = useTranslations("auth");
-  const { form, apiError, identifier, countdown, onSubmit, handleResend, handleUseDifferentIdentifier } = useRegisterOtpStep();
-  const otpValue = form.watch("code");
+  const otpStep = useRegisterOtpStep();
 
   return (
-    <>
-      <CardHeader className="relative space-y-1 text-center">
-        <OtpDevTotpFill
-          identifier={identifier}
-          setOtpCode={(code) => form.setValue("code", code, { shouldDirty: true, shouldValidate: true })}
-          onApiError={(err) => apiError.set(err ?? undefined)}
-        />
-        <AuthStepHeaderIcon icon={ShieldCheck} />
-        <CardTitle className="text-2xl">{t("otp.title")}</CardTitle>
-        <CardDescription className="text-center">
-          <span className="block">{t("otp.subtitle").trimEnd()}</span>
-          <span className="block font-semibold text-foreground">{identifier}</span>
-        </CardDescription>
-      </CardHeader>
-      <CardContent className="space-y-4">
-        <ErrorAlert message={apiError.message} />
-        <form id="register-otp-form" onSubmit={onSubmit} className="space-y-4">
-          <FieldGroup>
-            <Controller
-              name="code"
-              control={form.control}
-              render={({ field, fieldState }) => (
-                <Field data-invalid={fieldState.invalid}>
-                  <div className="flex justify-center">
-                    <InputOTP
-                      {...field}
-                      id="register-otp-code"
-                      maxLength={6}
-                      aria-invalid={fieldState.invalid}
-                      onComplete={() => {
-                        void onSubmit();
-                      }}
-                    >
-                      <InputOTPGroup className="*:data-[slot=input-otp-slot]:h-14 *:data-[slot=input-otp-slot]:w-13 *:data-[slot=input-otp-slot]:text-xl">
-                        {Array.from({ length: 6 }).map((_, i) => (
-                          <InputOTPSlot key={i} index={i} />
-                        ))}
-                      </InputOTPGroup>
-                    </InputOTP>
-                  </div>
-                  <FieldMessageSlot>
-                    {fieldState.invalid && fieldState.error?.message ? <FieldError errors={[{ message: t(fieldState.error.message) }]} /> : null}
-                  </FieldMessageSlot>
-                </Field>
-              )}
-            />
-          </FieldGroup>
-        </form>
-
-        <Button type="submit" form="register-otp-form" className="w-full" disabled={form.formState.isSubmitting || otpValue.length !== 6}>
-          {t("otp.verifyCode")}
-        </Button>
-
-        <div className="flex flex-col items-center gap-1 text-center text-sm text-muted-foreground">
-          <span>{t("otp.resend")}</span>
-          <AuthActionLink
-            label={countdown.isActive ? t("otp.resendIn", { time: formatCountdown(countdown.seconds) }) : t("otp.resendCode")}
-            onClick={handleResend}
-            disabled={countdown.isActive}
-            className="h-auto disabled:cursor-not-allowed disabled:opacity-50"
-          />
-        </div>
-
-        <div className="flex w-full flex-row items-center gap-2">
-          <Separator className="flex-1" />
-          <span className="text-xs uppercase tracking-wider text-muted-foreground">{t("otp.or")}</span>
-          <Separator className="flex-1" />
-        </div>
-
-        <div className="text-center">
-          <AuthActionLink href="/register?step=email" onClick={handleUseDifferentIdentifier} icon={MoveLeft} label={t("otp.useDifferentEmail")} />
-        </div>
-      </CardContent>
-    </>
+    <AuthOtpStep
+      formId="register-otp-form"
+      otpInputId="register-otp-code"
+      differentIdentifierHref="/register?step=email"
+      form={otpStep.form}
+      apiError={otpStep.apiError}
+      identifier={otpStep.identifier}
+      countdown={otpStep.countdown}
+      onSubmit={otpStep.onSubmit}
+      handleResend={otpStep.handleResend}
+      handleUseDifferentIdentifier={otpStep.handleUseDifferentIdentifier}
+    />
   );
 }
